+++ Makefile.am
@@ -376,6 +376,7 @@ libuv_la_SOURCES += src/unix/linux-core.c \
                     src/unix/linux-inotify.c \
                     src/unix/linux-syscalls.c \
                     src/unix/linux-syscalls.h \
+                    src/unix/procfs-exepath.c \
                     src/unix/proctitle.c \
                     src/unix/sysinfo-loadavg.c \
                     src/unix/sysinfo-memory.c
